refactor: organizing dashboard hooks
